source code for linked lists.
Doubly and Circular lists
o	Ex 6 – Gaddis modified: 
	Write a member function insertAtPos that inserts an element at a given position in a doubly linked list. Positions in the list start from 0. 
	Write a function that computes the sum of elements in a doubly linked list passed as parameter.
	Test your functions in the main where the user fills the list, we display then the elements and the sum of values.
o	Write a function that accepts a string as a parameter and determines whether the string is palindrome or not using a linked list or one of its variations. Test it in a main function where the user inputs a string as a parameter. The string could contain spaces.